BMC works until 1-3

+++ src/Bicycle.h
@@ -10,14 +10,14 @@
 
 #include "math.h"
 
-#define FPEDAL		(float)70
-#define RPEDAL		(float)0.015
-#define MU			(float)0.015
-#define RWHEEL		(float)0.622
-#define M			(float)70
-#define G			(float)9.81
-#define TS			(float) 1.0
-#define CINV		(float)0.8
+#define FPEDAL		(double)70
+#define RPEDAL		(double)0.015
+#define MU			(double)0.015
+#define RWHEEL		(double)0.622
+#define M			(double)70
+#define G			(double)9.81
+#define TS			(double) 1.0
+#define CINV		(double)0.8
 #define C1			(FPEDAL * ((RPEDAL) / (pow(RWHEEL, 2) * M)))
 #define C2			(-1 * MU * G) / RWHEEL
 
@@ -25,19 +25,19 @@
 #define DEVIATION 	3
 #define TIMEFRAMES 	200
 
-#define TEETHF1 	34
-#define TEETHF2 	50
+#define TEETHF1 	(double) 34
+#define TEETHF2 	(double) 50
 
-#define TEETHR1 	25
-#define TEETHR2 	23
-#define TEETHR3 	21
-#define TEETHR4 	19
-#define TEETHR5 	17
-#define TEETHR6 	16
-#define TEETHR7 	15
-#define TEETHR8 	14
-#define TEETHR9 	13
-#define TEETHR10 	12
+#define TEETHR1 	(double) 25
+#define TEETHR2 	(double) 23
+#define TEETHR3 	(double) 21
+#define TEETHR4 	(double) 19
+#define TEETHR5 	(double) 17
+#define TEETHR6 	(double) 16
+#define TEETHR7 	(double) 15
+#define TEETHR8 	(double) 14
+#define TEETHR9 	(double) 13
+#define TEETHR10 	(double) 12
 
 
 #endif /* BICYCLE_H_ */
